+++ main.go
@@ -5,12 +5,9 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/adjust/redismq"
-	"github.com/google/uuid"
 	"github.com/gorilla/mux"
-	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/mongo"
 	"html/template"
-	"io/ioutil"
 	"net/http"
 	"net/url"
 	"os"
@@ -18,6 +15,7 @@ import (
 	"receipt_collector/markets"
 	"receipt_collector/mongo_client"
 	"receipt_collector/nalogru_client"
+	"receipt_collector/receipts"
 	"receipt_collector/users"
 	"time"
 )
@@ -44,8 +42,8 @@ func main() {
 	router := mux.NewRouter()
 	router.HandleFunc("/api/market", markets.MarketsBaseHandler)
 	router.HandleFunc("/api/market/{id:[a-zA-Z0-9]+}", markets.ConcreteMarketHandler).Methods(http.MethodPut, http.MethodGet, http.MethodDelete)
-	router.HandleFunc("/api/receipt", getReceiptHandler)
-	router.HandleFunc("/api/receipt/from-bar-code", addReceiptHandler)
+	router.HandleFunc("/api/receipt", receipts.GetReceiptHandler).Methods(http.MethodGet)
+	router.HandleFunc("/api/receipt/from-bar-code", receipts.AddReceiptHandler).Methods(http.MethodPost)
 	loginRoute := "/api/login"
 	router.HandleFunc(loginRoute, users.LoginHandler).Methods(http.MethodPost)
 	registerUnauthenticatedRoutes(router)
@@ -60,90 +58,6 @@ func registerUnauthenticatedRoutes(router *mux.Router) {
 	router.HandleFunc(registrationRoute, users.UserRegistrationHandler)
 	http.Handle(registrationRoute, router)
 
-}
-
-func getReceiptHandler(writer http.ResponseWriter, request *http.Request) {
-	if request.Method != http.MethodGet {
-		writer.WriteHeader(http.StatusNotFound)
-		return
-	}
-	ctx, _ := context.WithTimeout(context.Background(), 10*time.Second)
-	defer func() {
-		err := request.Body.Close()
-		if err != nil {
-			fmt.Printf("error while request body close %s", err)
-		}
-	}()
-	client := mongo_client.GetMongoClient(mongoUrl, mongoUser, mongoSecret)
-	defer func() {
-		err := client.Disconnect(ctx)
-		if err != nil {
-			fmt.Printf("error while mongo disconnect %s", err)
-		}
-	}()
-	collection := client.Database("receipt_collection").Collection("receipts")
-	cursor, err := collection.Find(ctx, bson.D{})
-	if err != nil {
-		writer.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-	defer func() {
-		err := cursor.Close(ctx)
-		if err != nil {
-			fmt.Printf("error while cursor close %s", err)
-		}
-	}()
-	var receipts = readReceipts(cursor, ctx)
-	resp, err := json.Marshal(receipts)
-	if err != nil {
-		writer.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-	_, err = writer.Write(resp)
-	if err != nil {
-		writer.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-}
-
-func readReceipts(cursor *mongo.Cursor, context context.Context) []Receipt {
-	var receipts = make([]Receipt, 0, 0)
-	for cursor.Next(context) {
-		var receipt Receipt
-		err := cursor.Decode(&receipt)
-		check(err)
-		receipts = append(receipts, receipt)
-	}
-	return receipts
-}
-
-func addReceiptHandler(writer http.ResponseWriter, request *http.Request) {
-	if request.Method != http.MethodPost {
-		writer.WriteHeader(http.StatusNotFound)
-		return
-	}
-	defer func() {
-		err := request.Body.Close()
-		if err != nil {
-			fmt.Printf("error while request body close %s", err)
-		}
-	}()
-
-	queryString := request.URL.RawQuery
-	err := queueRequest(requestsQueue, queryString)
-	if err != nil {
-		fmt.Printf("error while queue request(%s). %s", queryString, err)
-		writer.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-}
-
-func dumpToFile(rawReceipt []byte) {
-	unique, _ := uuid.NewUUID()
-	fileName := dumpDirectory + unique.String() + ".json"
-	err := ioutil.WriteFile(fileName, rawReceipt, 0644)
-	fmt.Println(fileName)
-	check(err)
 }
 
 func saveResponse(queue *redismq.Queue, response []byte) {
@@ -172,17 +86,14 @@ func parseQuery(form *url.Values) nalogru_client.ParseResult {
 	}
 }
 
-func parseReceipt(bytes []byte) Receipt {
-	var receipt map[string]map[string]Receipt
+func parseReceipt(bytes []byte) receipts.Receipt {
+	var receipt map[string]map[string]receipts.Receipt
 	err := json.Unmarshal(bytes, &receipt)
 	check(err)
 
 	res := receipt["document"]["receipt"]
 
 	return res
-}
-func queueRequest(requestQueue *redismq.Queue, parsedBarCode string) error {
-	return requestQueue.Put(parsedBarCode)
 }
 
 func consumeRawReceipts(rawQueue *redismq.Queue) {
