+++ receipts/receipt.go
@@ -1,4 +1,4 @@
-package main
+package receipts
 
 import (
 	"fmt"
